refactor(scripts): start using new generators for logo image
