+++ src/app/common/images/misc-images.ts
@@ -1,0 +1,15 @@
+import { InjectionToken } from '@angular/core'
+import miscImages from '../../../data/images/misc.json'
+import { ImageAsset } from './image-asset'
+
+export const MISC_IMAGES = new InjectionToken<MiscImages>(
+  'Miscellaneous images',
+  {
+    factory: () => miscImages,
+  },
+)
+
+export interface MiscImages {
+  horizontalLogo: ImageAsset
+  aboutPortrait: ImageAsset
+}
