+++ src/app/logo/logo.component.html
@@ -1,10 +1,10 @@
 <!-- srcset subset from Angular defaults -->
 <a routerLink="/">
   <img
-    [ngSrc]="logoImages.horizontal.filePath"
-    [alt]="logoImages.horizontal.alt"
-    [width]="logoImages.horizontal.width"
-    [height]="logoImages.horizontal.height"
+    [ngSrc]="horizontalLogo.filePath"
+    [alt]="horizontalLogo.alt"
+    [width]="horizontalLogo.width"
+    [height]="horizontalLogo.height"
     [priority]="true"
     [ngSrcset]="srcSet"
     [sizes]="sizes"
