+++ src/app/project-page/project-images.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core'
-import { ImageAsset } from '../common/images/types'
+import { ImageAsset } from '../common/images/image-asset'
 import { JsonFetcher } from '../common/json-fetcher/json-fetcher'
 import { PROJECTS_DIR } from '../common/directories'
 import { from, Observable } from 'rxjs'
