+++ src/app/projects-page/project-item/project-item.ts
@@ -1,4 +1,4 @@
-import { ImageAsset } from '../../common/images/types'
+import { ImageAsset } from '../../common/images/image-asset'
 
 export interface Project {
   readonly slug: string
