--- src/data/images/logos/images.json
@@ -1,1 +1,0 @@
-[]
